Update README.md
Colour detection is the process of detecting the name of any color. This project can automatically get the name of the color by clicking on them. So for this, we will have a data file that contains the color name and its values. Then we will calculate the distance from each color and find the shortest one.